requirements.txt
flask
ultralytics
opencv-python-headless
numpy
openai
python-dotenv
Pillow